requirements.txt
Flask
Flask-SQLAlchemy
Flask-Praetorian
Flask-RESTful
Flask-Cors
Flask-Login
SQLAlchemy
pymysql
python-dotenv
requests
Werkzeug
pytest
cryptography 